requirements.txt
# AUTO-GENERATED — do not edit by hand.
# Regenerate with: make sync-requirements
# Source of truth: backend/pyproject.toml ([project].dependencies)
#
# Consumed by scripts/prepareproject.sh (BYOC build hook). Production
# builds (Dockerfile, Container Manager) install from backend/pyproject.toml
# directly; this file exists only to keep BYOC working until that flow can
# read pyproject.toml.
fastapi>=0.115
uvicorn[standard]>=0.34
python-arango>=8.1
pydantic>=2.10
pydantic-settings>=2.7
python-multipart>=0.0.18
langchain>=0.3
langchain-anthropic>=0.3
langchain-openai>=0.3
structlog>=24.4
httpx>=0.28
celery[redis]>=5.4
python-docx>=1.1
pymupdf>=1.25
tiktoken>=0.9
openai>=1.60
redis>=5.2
mcp>=1.0
langgraph>=0.2
rdflib>=7.0
arango-entity-resolution>=0.1
aiofiles>=24.1
PyJWT>=2.8
prometheus_client>=0.21
